Build pydata and ipython using python 3.5.1
I wasn't able to update guix python3 to use 3.5.1
so I rolled my own module. I discovered how use
(gnu build-system python) package-with-explicit-python
to handle changing the selected python for python build
system, and even better it recurses through the inputs
to adjust the python for the dependencies.

The packages are mostly to build jupyter notebook and bokeh.